+++ src/main/java/org/zalando/actuate/autoconfigure/failsafe/CircuitBreakersEndpointAutoConfiguration.java
@@ -1,4 +1,4 @@
-package org.zalando.failsafeactuator;
+package org.zalando.actuate.autoconfigure.failsafe;
 
 import net.jodah.failsafe.CircuitBreaker;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,6 +6,7 @@ import org.springframework.boot.actuate.autoconfigure.endpoint.condition.Conditi
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.zalando.actuate.failsafe.CircuitBreakersEndpoint;
 
 import javax.annotation.Nullable;
 import java.util.Collections;
