+++ src/main/java/org/zalando/actuate/autoconfigure/failsafe/package-info.java
@@ -1,4 +1,4 @@
 @ParametersAreNonnullByDefault
-package org.zalando.failsafeactuator;
+package org.zalando.actuate.autoconfigure.failsafe;
 
 import javax.annotation.ParametersAreNonnullByDefault;
